Binary Search problem for interviews
Given an array A of N integers, and a target value K. A is sorted in non-decreasing order. Write an efficient program to find the start index and the end index of K in A.

Try writing code on your own. Check specially for boundary conditions:
For example N = 5, k = 1, [A]=1 1 1 1 1